requirements.txt
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
idna==2.7
pkg-resources==0.0.0
pyparsing==2.2.0
python-dateutil==2.7.3
PyYAML==3.13
requests==2.19.1
ruamel.yaml==0.15.44
six==1.11.0
strictyaml==0.11.9
svgwrite==1.1.12
urllib3==1.23
